Files from computer(Homework 2 Galbraith)
This is two files, but they contain the same information. I had to code this in Jupyter notebook because I use the school computers during class. However, I am trying to get it so that I can use python on my computer like the school ones. This is my code for Homework 2 for Phys 4840. I put them in the notebook app so there is the text file that was copied and pasted.